Remove unnecessary utilities, empty line from apt-packages
Build failed due to an empty item in the apt item list.

As discussed earlier, convmv and recode are unnecessary since the
actions they perform are manual actions and should be run by the
developer on their host machine. The tools serve a purpose for
troubleshooting character encoding issues, which are quite rare.