Prevent double invocation of 'sc' (fixes double SSH exit issue)
The sudo section did not have an exit in it, so it lead to the sc
script being invoked twice. Instead of fixing it, remove is, since
the sudo shorthand is actually useless and no automation needs it.
We can during testing write 'sudo' ourselves if needed.

Also remove unnecessary exit calls as they might hide underlying
non-exit values.

Also add a nicely informative while-loop that ensures the 'vagrant up'
will not fail if the provision commands are run too early.