WIP: Upgrade kernel to 5.3 series from Ubuntu hwe packages
This does not help much, since the virtualbox-guest-utils and
virtualbox-guest-utils-hwe will anyway be at version 5.2 in Ubuntu 18.04.

Only upgrade to 20.04 would bring VirtualBox 6.1 guest utils:
https://packages.ubuntu.com/search?keywords=virtualbox-guest-utils&searchon=names&suite=all&section=all

The kernel in the hwe package that is backported from Ubuntu 19.10 for
Ubuntu 18.04 contains a VirtualBox 6.0 module. Need to wait until there
is a HWE kernel package with the Linux kernel 5.4 version (from Ubuntu
20.04) that is likely to have a VirtualBox 6.1 module, but doing full
image upgrade is much better since then also the utils package would
be newer.