Bugfix: Touch the config.yml file to pass wait-mount
Maybe due to some Docker upgrade, the service now failed to start since
it was stuck on waiting for the condition in delay.conf to be met.

In a provision situation it will never be met, so simply bypass the issue
by touching the file that is expected to be seen.