Bugfix: Move the delay.conf to correct systemd service
The thing we actually want to prevent during provisioning is the Docker
container starting, while having the Docker service start is perfectly
fine and in fact necessary to be able to run 'docker pull'.

Having the delay.conf on the wrong service in the first place was a bug.
This fixes it and also cleans away to workaround used earlier to mitigate
this.

This should fix issues where the development environment has the domain
'seravo-wordpress', which was inherited from a Docker container running
(erroneously) during the Vagrant box provisioning.

Also the workaround might have left stray config.yml files with '1' as
contents around, so replacing the workaround with a proper bugfix also
fixes those symptoms.