Disable serial output during boot
This fixes compatibility issues with Virtualbox 6. The issues were
caused by the base image logging to a serial port.

Note: The serial removal is conditional, because it would otherwise fail CI.

Note: This affects only the boots, that are run after the build is done,
this means that the UART is needed during the build.