Fix error "cannot create directory ‘/data/log/vagrant’"
Occasionally it was noted that 'vagrant up' failed to run the command
wp-development-up with error message:

  ==> wordpress-box: Running trigger...
  mkdir: cannot create directory ‘/data/log/vagrant’: Permission denied
  Connection to 127.0.0.1 closed.

This was due to a race condition of the wp-wrapper-interactive running
before Docker had enough time to start up and create /data/log.

Add an extra check/delay for /data/log before proceeding to prevent this.